requirements.txt
Flask
flask-cors
python-dotenv
langchain-core
langchain-groq
